requirements.txt
streamlit
pillow
opencv-python-headless
pytesseract
numpy